+++ src/RcppExports.cpp
@@ -6,22 +6,23 @@
 using namespace Rcpp;
 
 // recordSwap_cpp
-std::vector< std::vector<int> > recordSwap_cpp(std::vector< std::vector<int> > data, Rcpp::List similar_cpp, std::vector<int> hierarchy, std::vector<int> risk_variables, int hid, int k_anonymity, double swaprate, double risk_threshold, std::vector<std::vector<double>> risk, int seed);
-RcppExport SEXP _recordSwapping_recordSwap_cpp(SEXP dataSEXP, SEXP similar_cppSEXP, SEXP hierarchySEXP, SEXP risk_variablesSEXP, SEXP hidSEXP, SEXP k_anonymitySEXP, SEXP swaprateSEXP, SEXP risk_thresholdSEXP, SEXP riskSEXP, SEXP seedSEXP) {
+std::vector< std::vector<int> > recordSwap_cpp(std::vector< std::vector<int> > data, int hid, std::vector<int> hierarchy, Rcpp::List similar_cpp, double swaprate, std::vector<std::vector<double>> risk, double risk_threshold, int k_anonymity, std::vector<int> risk_variables, std::vector<int> carry_along, int seed);
+RcppExport SEXP _recordSwapping_recordSwap_cpp(SEXP dataSEXP, SEXP hidSEXP, SEXP hierarchySEXP, SEXP similar_cppSEXP, SEXP swaprateSEXP, SEXP riskSEXP, SEXP risk_thresholdSEXP, SEXP k_anonymitySEXP, SEXP risk_variablesSEXP, SEXP carry_alongSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< std::vector< std::vector<int> > >::type data(dataSEXP);
-    Rcpp::traits::input_parameter< Rcpp::List >::type similar_cpp(similar_cppSEXP);
-    Rcpp::traits::input_parameter< std::vector<int> >::type hierarchy(hierarchySEXP);
-    Rcpp::traits::input_parameter< std::vector<int> >::type risk_variables(risk_variablesSEXP);
     Rcpp::traits::input_parameter< int >::type hid(hidSEXP);
-    Rcpp::traits::input_parameter< int >::type k_anonymity(k_anonymitySEXP);
+    Rcpp::traits::input_parameter< std::vector<int> >::type hierarchy(hierarchySEXP);
+    Rcpp::traits::input_parameter< Rcpp::List >::type similar_cpp(similar_cppSEXP);
     Rcpp::traits::input_parameter< double >::type swaprate(swaprateSEXP);
-    Rcpp::traits::input_parameter< double >::type risk_threshold(risk_thresholdSEXP);
     Rcpp::traits::input_parameter< std::vector<std::vector<double>> >::type risk(riskSEXP);
+    Rcpp::traits::input_parameter< double >::type risk_threshold(risk_thresholdSEXP);
+    Rcpp::traits::input_parameter< int >::type k_anonymity(k_anonymitySEXP);
+    Rcpp::traits::input_parameter< std::vector<int> >::type risk_variables(risk_variablesSEXP);
+    Rcpp::traits::input_parameter< std::vector<int> >::type carry_along(carry_alongSEXP);
     Rcpp::traits::input_parameter< int >::type seed(seedSEXP);
-    rcpp_result_gen = Rcpp::wrap(recordSwap_cpp(data, similar_cpp, hierarchy, risk_variables, hid, k_anonymity, swaprate, risk_threshold, risk, seed));
+    rcpp_result_gen = Rcpp::wrap(recordSwap_cpp(data, hid, hierarchy, similar_cpp, swaprate, risk, risk_threshold, k_anonymity, risk_variables, carry_along, seed));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -183,7 +184,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_recordSwapping_recordSwap_cpp", (DL_FUNC) &_recordSwapping_recordSwap_cpp, 10},
+    {"_recordSwapping_recordSwap_cpp", (DL_FUNC) &_recordSwapping_recordSwap_cpp, 11},
     {"_recordSwapping_setLevels_cpp", (DL_FUNC) &_recordSwapping_setLevels_cpp, 2},
     {"_recordSwapping_orderData_cpp", (DL_FUNC) &_recordSwapping_orderData_cpp, 2},
     {"_recordSwapping_setRisk_cpp", (DL_FUNC) &_recordSwapping_setRisk_cpp, 4},